+++ packages/skirmish/src/GameContext.res.mjs
@@ -14,6 +14,8 @@ let k = Kaplay({
   crisp: true
 });
 
+k.debug.inspect = true;
+
 let gameWidth = 360;
 
 let gameHeight = 649;